requirements.txt
click==6.7
sphinx_rtd_theme==0.2.4
setuptools==20.7.0
svgwrite==1.1.10
Pillow==4.2.1
